refactor: split workflow run analysis helpers
